requirements.txt
anyio==4.8.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.95.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==1.10.21
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.22.0
